requirements.txt
yfinance>=0.2.32
pandas-ta>=0.3.14b0
vaderSentiment>=3.3.2
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.25.0
groq>=0.4.2
pandas>=2.1.0
numpy>=1.24.0
